+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 
 - Fix options of some mods (dmitmel tweak pack) not being announced
+- Fix aim bounce turning on when entering the analysis menu
 
 ## [0.6.0] 2024-04-20